fix: Bazel invokes credential_helper as "<path> get"; expose hidden root get cmd
Bazel spawns `--credential_helper=<path>` binary with `get` as argv[1] per
the EngFlow Credential Helper spec. The CLI root had no `get` subcommand,
so every remote-cache/BES auth failed with `unknown command "get"`.

- Add hidden root `get` subcommand that forwards to bazelcredhelper.Run.
- Drop the now-unreachable `bazel credential-helper` cobra command and the
  bazel parent group (its remaining leaf commands register on
  common.ActivateCmd / common.EnableForCmd, not the group).
- Skip the version check for `get` — Bazel spawns the helper on every
  action; stderr chatter would flood build logs.
- Refresh stale comments in bazelrc template + template-inventory.
