+++ internal/config/bazel/bazel_template_inventory.go
@@ -20,7 +20,8 @@ type CommonTemplateInventory struct {
 	Timestamps   bool
 	// CLIPath is the absolute path of the bitrise-build-cache binary. On local
 	// dev (empty CIProvider) it drives `build --credential_helper=<CLIPath>`
-	// so the auth token is resolved per-build via `bazel credential-helper`
+	// so the auth token is resolved per-build via the hidden `get` subcommand
+	// (Bazel invokes `<CLIPath> get` per the EngFlow credential-helper spec)
 	// instead of being written literally into `~/.bazelrc`. Empty on CI —
 	// the CI branch keeps the literal `Bearer <token>` header for perf.
 	CLIPath      string
