+++ main.go
@@ -9,6 +9,7 @@ import (
 	_ "github.com/bitrise-io/bitrise-build-cache-cli/v3/cmd/daemon"
 	_ "github.com/bitrise-io/bitrise-build-cache-cli/v3/cmd/doctor"
 	_ "github.com/bitrise-io/bitrise-build-cache-cli/v3/cmd/file"
+	_ "github.com/bitrise-io/bitrise-build-cache-cli/v3/cmd/get"
 	_ "github.com/bitrise-io/bitrise-build-cache-cli/v3/cmd/gradle"
 	_ "github.com/bitrise-io/bitrise-build-cache-cli/v3/cmd/reactnative"
 	_ "github.com/bitrise-io/bitrise-build-cache-cli/v3/cmd/update"
